[FLINK-31500][table-planner] Move SqlAlterTableSchema conversion logic to SqlAlterTableSchemaConverter
