+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/operations/converters/SqlNodeConverters.java
@@ -50,6 +50,8 @@ public class SqlNodeConverters {
         register(new SqlTruncateTableConverter());
         register(new SqlShowFunctionsConverter());
         register(new SqlShowProcedureConverter());
+        register(new SqlAlterTableAddConverter());
+        register(new SqlAlterTableModifyConverter());
     }
 
     /**
